+++ Sources/Sentry/SentryDataCategoryMapper.m
@@ -59,8 +59,24 @@ SentryDataCategoryMapper ()
     if (value == kSentryDataCategoryAttachment) {
         category = kSentryDataCategoryAttachment;
     }
+    if (value == kSentryDataCategoryUserFeedback) {
+        category = kSentryDataCategoryUserFeedback;
+    }
     if (value == kSentryDataCategoryUnknown) {
         category = kSentryDataCategoryUnknown;
+    }
+
+    return category;
+}
+
++ (SentryDataCategory)mapStringToCategory:(NSString *)value
+{
+    SentryDataCategory category = kSentryDataCategoryUnknown;
+
+    for (int i = 0; i <= kSentryDataCategoryUnknown; i++) {
+        if ([value isEqualToString:SentryDataCategoryNames[i]]) {
+            return [SentryDataCategoryMapper mapIntegerToCategory:i];
+        }
     }
 
     return category;