+++ Sources/Sentry/SentryOptions.m
@@ -66,6 +66,7 @@ SentryOptions ()
 #if SENTRY_TARGET_PROFILING_SUPPORTED
         self.enableProfiling = NO;
 #endif
+        self.sendClientReports = YES;
 
         // Use the name of the bundle’s executable file as inAppInclude, so SentryInAppLogic
         // marks frames coming from there as inApp. With this approach, the SDK marks public
@@ -272,6 +273,9 @@ SentryOptions ()
     [self setBool:options[@"enableProfiling"]
             block:^(BOOL value) { self->_enableProfiling = value; }];
 #endif
+
+    [self setBool:options[@"sendClientReports"]
+            block:^(BOOL value) { self->_sendClientReports = value; }];
 
     if (nil != error && nil != *error) {
         return NO;