+++ Tests/SentryTests/Networking/RateLimits/TestEnvelopeRateLimitDelegate.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+class TestEnvelopeRateLimitDelegate: NSObject, SentryEnvelopeRateLimitDelegate {
+    
+    var envelopeItemsDropped = Invocations<SentryDataCategory>()
+    func envelopeItemDropped(_ dataCategory: SentryDataCategory) {
+        envelopeItemsDropped.record(dataCategory)
+    }
+}