+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat: Add Client Reports (#1733)
 - fix: enableProfiling option via initWithDict (#1743)
 
 ## 7.12.0